Moved CreateConfigFlagsWithRestConfig to kueuectl test util.
